requirements.txt
# FastAPI and server
fastapi==0.104.1
uvicorn[standard]==0.24.0
pydantic==2.5.0
pydantic-settings==2.1.0

# WebSocket
# websockets==12.0

# Database
sqlalchemy==2.0.23

# Utilities
python-multipart==0.0.6
python-dotenv==1.0.0

# Socket.IO (for socketio_app)
python-socketio[asyncio]==5.11.0
